fix(yfm): fix circular deps
Made-with: Cursor
