+++ packages/editor/package.json
@@ -26,7 +26,7 @@
     "test:cov": "jest --coverage",
     "test:watch": "jest --watchAll",
     "test:esbuild": "node tests/esbuild-test/esbuild-tester.js",
-    "test:circular-deps": "node scripts/check-circular-deps.js 7",
+    "test:circular-deps": "node scripts/check-circular-deps.js 4",
     "prepack": "cp ../../README.md ./README.md",
     "postpack": "rm -f ./README.md",
     "prepublishOnly": "pnpm run lint && pnpm run clean && pnpm run build"
